Adds ProjectE Quests
I was originally going to have this chapter be "good" but considering there aren't many MTO updates and this quest mod is annoying this is as far as I got. It still goes over the main parts of the mod, but that's about it. I made these a while ago and forgot to commit them, so here they are.